Splitted PopupIndicators.
